implement util function to get token from activity
